finish community base function

+++ src/main/resources/db/migration/V1__Create_user_table.sql
@@ -1,6 +1,6 @@
 create table USER
 (
-	ID int AUTO_INCREMENT primary key,
+	ID bigint AUTO_INCREMENT primary key,
 	ACCOUNT_ID VARCHAR(100),
 	NAME VARCHAR(50),
 	TOKEN CHAR(36),
