+++ src/main/resources/db/migration/V3__Create_question_table.sql
@@ -1,11 +1,11 @@
 CREATE TABLE question
 (
-    id int AUTO_INCREMENT PRIMARY KEY,
+    id bigint AUTO_INCREMENT PRIMARY KEY,
     title varchar(50),
     description text,
     gmt_create long,
     gmt_modified long,
-    creator int,
+    creator bigint,
     comment_count int,
     view_count int,
     like_count int,
