+++ src/main/resources/db/migration/V6__Create_comment_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE comment
     id bigint AUTO_INCREMENT PRIMARY KEY,
     parent_id bigint NOT NULL,
     type int NOT NULL,
-    commentor int NOT NULL,
+    commentor bigint NOT NULL,
     content varchar(1024),
     gmt_create bigint NOT NULL,
     gmt_modified bigint NOT NULL,
